revert memry.c
