+++ .vscode/tasks.json
@@ -33,10 +33,13 @@
             "command": "gcc",
             "args": [
                 "build/test/${fileBasenameNoExtension}.o",
-                "build/array.o",
+                "build/arrays.o",
+                "build/collections.o",
+                "build/farray.o",
                 "build/list.o",
-                "build/slotarray.o",
                 "build/memory.o",
+                "build/parray.o",
+                "build/slotarray.o",
                 "-o",
                 "build/test/${fileBasenameNoExtension}",
                 "-lstest",
