+++ logs/test_mempaging.log
@@ -1,0 +1,3 @@
+[1] core_mempaging_set       :   3 :         2025-12-12  19:05:36
+=================================================================
+Running: Cross-page tracking                    2978.716 us  [FAIL]
