+++ src/memory.c
@@ -65,10 +65,10 @@ static slotarray memory_slots = NULL;
 static bool memory_ready = false;
 
 // Static buffers to avoid bootstrap allocations
-#define TRACKER_CAPACITY 4096
+#define TRACKER_CAPACITY 512
 #define PAGE_ARRAY_CAPACITY 16
-#define PAGE_SLOTS_CAPACITY 4096
-#define MAX_PAGES 16
+#define PAGE_SLOTS_CAPACITY 512
+// #define MAX_PAGES 16
 
 // Pre-allocated buffer for the slotarray used as the memory tracker.
 // This avoids dynamic allocation during init, preventing self-tracking issues.
@@ -86,7 +86,7 @@ static struct sc_pointer_array memory_pages_struct = {
     .end = (addr)(page_bucket + PAGE_ARRAY_CAPACITY)};
 
 // Pre-allocated page structures
-static struct memory_page memory_pages_pool[MAX_PAGES];
+// static struct memory_page memory_pages_pool[MAX_PAGES];
 
 // The current memory page, encapsulating the tracker slotarray and free head index.
 static struct memory_page current_page_struct = {
@@ -95,41 +95,41 @@ static struct memory_page current_page_struct = {
     .next = NULL,
     .prev = NULL};
 
-static usize next_page_index = 1; // Next available page in pool
-static usize page_count = 1;      // Number of active pages
+// static usize next_page_index = 1; // Next available page in pool
+// static usize page_count = 1;      // Number of active pages
 
 // Create a new memory page with dynamically allocated slotarray buffer
-static struct memory_page *create_new_page(void) {
-   if (next_page_index >= MAX_PAGES) {
-      return NULL; // No more pages available
-   }
-   struct memory_page *new_page = &memory_pages_pool[next_page_index++];
-   page_count++;
-   // Allocate buffer for slotarray using raw malloc (not tracked)
-   addr *buffer = malloc(PAGE_SLOTS_CAPACITY * sizeof(addr));
-   if (!buffer) {
-      return NULL;
-   }
-   // Allocate slotarray struct
-   struct sc_slotarray *new_slots = malloc(sizeof(struct sc_slotarray));
-   if (!new_slots) {
-      free(buffer);
-      return NULL;
-   }
-   new_slots->array.buffer = buffer;
-   new_slots->array.end = buffer + PAGE_SLOTS_CAPACITY;
-   new_slots->stride = sizeof(addr);
-   new_slots->can_grow = true;
-   // Initialize all to ADDR_EMPTY
-   for (usize i = 0; i < PAGE_SLOTS_CAPACITY; i++) {
-      buffer[i] = ADDR_EMPTY;
-   }
-   new_page->slots = new_slots;
-   new_page->free_head = (struct index){0, 0};
-   new_page->next = NULL;
-   new_page->prev = current_page;
-   return new_page;
-}
+// static struct memory_page *create_new_page(void) {
+//    if (next_page_index >= MAX_PAGES) {
+//       return NULL; // No more pages available
+//    }
+//    struct memory_page *new_page = &memory_pages_pool[next_page_index++];
+//    page_count++;
+//    // Allocate buffer for slotarray using raw malloc (not tracked)
+//    addr *buffer = malloc(PAGE_SLOTS_CAPACITY * sizeof(addr));
+//    if (!buffer) {
+//       return NULL;
+//    }
+//    // Allocate slotarray struct
+//    struct sc_slotarray *new_slots = malloc(sizeof(struct sc_slotarray));
+//    if (!new_slots) {
+//       free(buffer);
+//       return NULL;
+//    }
+//    new_slots->array.buffer = buffer;
+//    new_slots->array.end = buffer + PAGE_SLOTS_CAPACITY;
+//    new_slots->stride = sizeof(addr);
+//    new_slots->can_grow = true;
+//    // Initialize all to ADDR_EMPTY
+//    for (usize i = 0; i < PAGE_SLOTS_CAPACITY; i++) {
+//       buffer[i] = ADDR_EMPTY;
+//    }
+//    new_page->slots = new_slots;
+//    new_page->free_head = (struct index){0, 0};
+//    new_page->next = NULL;
+//    new_page->prev = current_page;
+//    return new_page;
+// }
 
 // allocate a block of memory of the specified size
 static object memory_alloc(usize size, bool zee) {
@@ -141,27 +141,9 @@ static object memory_alloc(usize size, bool zee) {
       // Try to track the allocation in the current page
       int result = SlotArray.add(current_page->slots, ptr);
       if (result == ERR) {
-         // Current page is full, create a new page
-         struct memory_page *new_page = create_new_page();
-         if (new_page) {
-            // Link the new page
-            current_page->next = new_page;
-            // Add to pages array
-            // PArray.set(memory_pages, page_count++, (addr)new_page);  // Temporarily disabled
-            // Switch to new page
-            current_page = new_page;
-            // Now try to add to the new page
-            result = SlotArray.add(current_page->slots, ptr);
-            if (result == ERR) {
-               // New page also failed, free and return NULL
-               free(ptr);
-               return NULL;
-            }
-         } else {
-            // No more pages, free and return NULL
-            free(ptr);
-            return NULL;
-         }
+         // Current page is full, cannot allocate more
+         free(ptr);
+         return NULL;
       }
    }
    return ptr;
@@ -169,21 +151,16 @@ static object memory_alloc(usize size, bool zee) {
 
 // dispose of a previously allocated block of memory
 static void memory_dispose(object ptr) {
-   if (memory_ready) {
-      // Traverse all pages to find and remove from tracker
-      struct memory_page *page = &current_page_struct; // Start from first page
-      while (page) {
-         usize cap = SlotArray.capacity(page->slots);
-         for (usize i = 0; i < cap; i++) {
-            if (!SlotArray.is_empty_slot(page->slots, i)) {
-               object val;
-               if (SlotArray.get_at(page->slots, i, &val) == 0 && val == ptr) {
-                  SlotArray.remove_at(page->slots, i);
-                  goto found;
-               }
+   if (memory_ready && current_page) {
+      usize cap = SlotArray.capacity(current_page->slots);
+      for (usize i = 0; i < cap; i++) {
+         if (!SlotArray.is_empty_slot(current_page->slots, i)) {
+            object val;
+            if (SlotArray.get_at(current_page->slots, i, &val) == 0 && val == ptr) {
+               SlotArray.remove_at(current_page->slots, i);
+               goto found;
             }
          }
-         page = page->next;
       }
    found:;
    }
@@ -192,20 +169,15 @@ static void memory_dispose(object ptr) {
 
 // check if a given memory pointer is currently being tracked
 static bool memory_is_tracking(object ptr) {
-   if (!memory_ready)
+   if (!memory_ready || !current_page)
       return false;
-   // Traverse all pages
-   struct memory_page *page = &current_page_struct;
-   while (page) {
-      usize cap = SlotArray.capacity(page->slots);
-      for (usize i = 0; i < cap; i++) {
-         if (!SlotArray.is_empty_slot(page->slots, i)) {
-            object val;
-            if (SlotArray.get_at(page->slots, i, &val) == 0 && val == ptr)
-               return true;
-         }
+   usize cap = SlotArray.capacity(current_page->slots);
+   for (usize i = 0; i < cap; i++) {
+      if (!SlotArray.is_empty_slot(current_page->slots, i)) {
+         object val;
+         if (SlotArray.get_at(current_page->slots, i, &val) == 0 && val == ptr)
+            return true;
       }
-      page = page->next;
    }
    return false;
 }
@@ -213,63 +185,42 @@ static bool memory_is_tracking(object ptr) {
 // track a memory pointer
 static void memory_track(object ptr) {
    if (memory_ready && current_page) {
-      int result = SlotArray.add(current_page->slots, ptr);
-      if (result == ERR) {
-         // Current page full, create new page
-         struct memory_page *new_page = create_new_page();
-         if (new_page) {
-            current_page->next = new_page;
-            PArray.set(memory_pages, page_count++, (addr)new_page);
-            current_page = new_page;
-            SlotArray.add(current_page->slots, ptr);
-         }
-         // If failed, silently ignore
-      }
+      SlotArray.add(current_page->slots, ptr);
    }
 }
 
 // untrack a memory pointer
 static void memory_untrack(object ptr) {
-   if (memory_ready) {
-      // Traverse all pages
-      struct memory_page *page = &current_page_struct;
-      while (page) {
-         usize cap = SlotArray.capacity(page->slots);
-         for (usize i = 0; i < cap; i++) {
-            if (!SlotArray.is_empty_slot(page->slots, i)) {
-               object val;
-               if (SlotArray.get_at(page->slots, i, &val) == 0 && val == ptr) {
-                  SlotArray.remove_at(page->slots, i);
-                  return;
-               }
+   if (memory_ready && current_page) {
+      usize cap = SlotArray.capacity(current_page->slots);
+      for (usize i = 0; i < cap; i++) {
+         if (!SlotArray.is_empty_slot(current_page->slots, i)) {
+            object val;
+            if (SlotArray.get_at(current_page->slots, i, &val) == 0 && val == ptr) {
+               SlotArray.remove_at(current_page->slots, i);
+               return;
             }
          }
-         page = page->next;
       }
    }
 }
 
 // internal function to swap tracked pointers (for realloc optimization)
 static bool memory_swap_tracked_pointers(object old_ptr, object new_ptr) {
-   if (!memory_ready || !old_ptr || !new_ptr)
+   if (!memory_ready || !current_page || !old_ptr || !new_ptr)
       return false;
 
-   // Traverse all pages to find and swap the pointer
-   struct memory_page *page = &current_page_struct;
-   while (page) {
-      usize cap = SlotArray.capacity(page->slots);
-      for (usize i = 0; i < cap; i++) {
-         if (!SlotArray.is_empty_slot(page->slots, i)) {
-            object val;
-            if (SlotArray.get_at(page->slots, i, &val) == 0 && val == old_ptr) {
-               // Found the old pointer, replace it with new pointer
-               addr *slot = (addr *)((char *)page->slots->array.buffer + i * page->slots->stride);
-               *slot = (addr)new_ptr;
-               return true;
-            }
+   usize cap = SlotArray.capacity(current_page->slots);
+   for (usize i = 0; i < cap; i++) {
+      if (!SlotArray.is_empty_slot(current_page->slots, i)) {
+         object val;
+         if (SlotArray.get_at(current_page->slots, i, &val) == 0 && val == old_ptr) {
+            // Found the old pointer, replace it with new pointer
+            addr *slot = (addr *)((char *)current_page->slots->array.buffer + i * current_page->slots->stride);
+            *slot = (addr)new_ptr;
+            return true;
          }
       }
-      page = page->next;
    }
    return false; // old_ptr not found
 }
@@ -330,34 +281,23 @@ static void memory_teardown(void) {
    if (!memory_ready)
       return;
 
-   // Dispose all tracked allocations from all pages
-   struct memory_page *page = &current_page_struct;
-   while (page) {
-      if (page->slots) {
-         usize cap = SlotArray.capacity(page->slots);
-         for (usize i = 0; i < cap; i++) {
-            if (!SlotArray.is_empty_slot(page->slots, i)) {
-               object val;
-               if (SlotArray.get_at(page->slots, i, &val) == 0) {
-                  free(val); // raw free, bypass dispose to avoid recursion
-               }
+   // Dispose all tracked allocations from the current page
+   if (current_page && current_page->slots) {
+      usize cap = SlotArray.capacity(current_page->slots);
+      for (usize i = 0; i < cap; i++) {
+         if (!SlotArray.is_empty_slot(current_page->slots, i)) {
+            object val;
+            if (SlotArray.get_at(current_page->slots, i, &val) == 0) {
+               free(val); // raw free, bypass dispose to avoid recursion
             }
          }
-         // Free the buffer and struct if not the static ones
-         if (page->slots != &memory_slots_struct) {
-            free(page->slots->array.buffer);
-            free(page->slots);
-         }
       }
-      page = page->next;
    }
 
    // Reset globals
    memory_slots = NULL;
    current_page = NULL;
    memory_pages = NULL;
-   next_page_index = 1;
-   page_count = 1;
    memory_ready = false;
 }
 
@@ -383,5 +323,5 @@ slotarray Memory_get_tracker(void) {
 }
 
 usize Memory_get_page_count(void) {
-   return page_count;
+   return 1;
 }
